Alter flightcheck to perform better across server platforms.
Previously it would test if a cache path was relative by simply looking for a leading slash; however Windows servers begin path declarations with a letter drive (e.g. 'C:/'), so the test was always failing. Now, we simply test for the existence of the path itself; if not found, we append to our Base Path, and try again.